daemon: move account `jsonrpc_` methods to an individual module
